Cohomology

GKMtest.isGKMclass — Function

isGKMclass(c::FreeModElem{QQMPolyRingElem}, G::AbstractGKM_graph) -> Bool

Return true if the given class is an actual GKM cohomology class. This holds if and only if the difference between localizations at fixed points connected through an edge e is divisible by the weight of e.

isGKMclass(c::FreeModElem{QQMPolyRingElem}, R::GKM_cohomology_ring) -> Bool

Return true if the given class is an actual GKM cohomology class. This holds if and only if the difference between localizations at fixed points connected through an edge e is divisible by the weight of e.

GKMtest.pointClass — Function

pointClass(vertexLabel::String, G::AbstractGKM_graph) -> FreeModElem{QQMPolyRingElem}

Return the equivariant Poincare dual of the fixed point with given label

pointClass(vertexLabel::String, R::GKM_cohomology_ring) -> FreeModElem{QQMPolyRingElem}

Return the equivariant Poincare dual of the fixed point with given label

pointClass(vertex::Int, G::AbstractGKM_graph) -> FreeModElem{QQMPolyRingElem}

Return the equivariant Poincare dual of the given fixed point

pointClass(vertex::Int, R::GKM_cohomology_ring) -> FreeModElem{QQMPolyRingElem}

Return the equivariant Poincare dual of the given fixed point

GKMtest.PDClass — Function

PDClass(gkmSub::AbstractGKM_subgraph) -> FreeModElem{QQMPolyRingElem}

Return the equivariant Poincare dual cohomology class of the GKM subgraph.

GKMtest.weightClass — Function

weightClass(e::Edge, G::AbstractGKM_graph) -> QQMPolyRingElem

Return the weight of the edge e as an element of the coefficient ring of the equivariant cohomology theory.

weightClass(e::Edge, R::GKM_cohomology_ring) -> QQMPolyRingElem

Return the weight of the edge e as an element of the coefficient ring of the equivariant cohomology theory.

GKMtest.eulerClass — Function

eulerClass(vertex::Int, G::AbstractGKM_graph) -> QQMPolyRingElem

Return the euler class of the normal bundle of the fixed point v.

eulerClass(vertex::Int, R::GKM_cohomology_ring) -> QQMPolyRingElem

Return the euler class of the normal bundle of the fixed point v.

eulerClass(vertexLabel::String, G::AbstractGKM_graph) -> QQMPolyRingElem

Return the euler class of the normal bundle of the fixed point with the given label.

eulerClass(vertexLabel::String, R::GKM_cohomology_ring) -> QQMPolyRingElem

Return the euler class of the normal bundle of the fixed point with the given label.

GKMtest.integrateGKMClass — Function

integrateGKMClass(class::FreeModElem{QQMPolyRingElem}, G::AbstractGKM_graph; check::Bool=true) -> QQMPolyRingElem

Integrate the GKM class, yielding an element of the coefficient ring. This only works if isGKMclass(class,R) == true.

integrateGKMClass(class::FreeModElem{QQMPolyRingElem}, R::GKM_cohomology_ring; check::Bool=true) -> QQMPolyRingElem

Integrate the GKM class, yielding an element of the coefficient ring. This only works if isGKMclass(class,R) == true.

GKMtest.firstChernClass — Function

firstChernClass(G::AbstractGKM_graph) -> FreeModElem{QQMPolyRingElem}

Return the equivariant first chern class of the GKM space.

firstChernClass(R::GKM_cohomology_ring) -> FreeModElem{QQMPolyRingElem}

Return the equivariant first chern class of the GKM space of which R is the cohomology ring.

GKMtest.integrateOverEdge — Function

integrateOverEdge(class::FreeModElem{QQMPolyRingElem}, G::AbstractGKM_graph, e::Edge) -> AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}

Integrate the cohomology class over the curve represented by the GKM graph edge.

integrateOverEdge(class::FreeModElem{QQMPolyRingElem}, R::GKM_cohomology_ring, e::Edge) -> AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}

Integrate the cohomology class over the curve represented by the GKM graph edge.

GKMtest.integrateClass — Function

integrateClass(class::FreeModElem{QQMPolyRingElem}, G::AbstractGKM_graph) -> AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}

Integrate the cohomology class, yielding an element of the fraction field of the equivariant coefficient ring. If isGKMclass(class, R) == true, this fraction will represent a ring element. Use integrateGKMClass() instead if you know that the class is a GKM class.

integrateClass(class::FreeModElem{QQMPolyRingElem}, R::GKM_cohomology_ring) -> AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}

Integrate the cohomology class, yielding an element of the fraction field of the equivariant coefficient ring. If isGKMclass(class, R) == true, this fraction will represent a ring element. Use integrateGKMClass() instead if you know that the class is a GKM class.