Connections

Functions regarding connections.

GKMtest.get_GKM_connection — Function

get_GKM_connection(gkm::AbstractGKM_graph) -> Union{Nothing, GKM_connection}

It returns the GKM_connection of the given GKM graph if it is unique or has been set manually. If it is unique and hasn't been calculated, it is saved in the gkm object. If the connection is not unique and hasn't been defined manually, it returns nothing.

GKMtest.set_GKM_connection! — Function

set_GKM_connection!(gkm::AbstractGKM_graph)

It finds, if possible, a GKM connection of G, and save it in the gkm object.

GKMtest.GKM_isValidConnection — Function

GKM_isValidConnection(con::GKM_connection; printDiagnostics::Bool=true) -> Bool

It returns true if the given GKM connection is valid. This holds if and only if all of the following hold:

con and a are set for all (Edge(v,w), Edge(v,u)) where vw and vu are connected in the graph

con maps every (e,e) to reverse(e)

a maps every (e,e) to 2

Every pair (e,ei) with same source satisfies the relation of the associated a's, i.e. w[ei'] = w[ei] - a[(e,ei)] * w[e]