These are the main properties of GKM graphs.

Oscar.betti_numbers — Function

betti_numbers(G::AbstractGKM_graph) -> Vector{Int64}

Return the array betti such that betti_numbers[i+1] is the 2i-th (combinatorial) Betti number for i from 0 to the valency of G, as defined in [GZ01, section 1.3].

Note

i ranges from 0 to the valency of G, that can be obtained by valency(G).

from [GZ01, Theorem 1.3.2], the combinatorial Betti numbers equal the Betti numbers of the underlying GKM space if the torus action is Hamiltonian.

Warning

betti_numbers[1] is the 0-th Betti number, since Julia arrays are 1-based and not 0-based.

GKMtest.valency — Function

valency(G::AbstractGKM_graph) -> Int64

It returns the valency of G.

GKMtest.rank_torus — Function

rank_torus(G::AbstractGKM_graph) -> Int64

It returns the rank of the torus acting on G. That is, the rank of the character group.

GKMtest.is2_indep — Function

is2_indep(G::AbstractGKM_graph) -> Bool

It returns true or false depending if G is 2-independent.

GKMtest.is3_indep — Function

is3_indep(G::AbstractGKM_graph) -> Bool

It returns true or false depending if G is 3-independent.