Constructors

These functions allow the construction of GKM varieties.

GKMtest.gkm_graph — Function

gkm_graph(g, labels, M, w; check=true, checkLabels=true) -> AbstractGKM_graph

Create a GKM graph from the given data.

Arguments

g::Graph: A OSCAR graph.

labels::Vector{String}: A vector of strings, used to denote the vertices.

M::AbstractAlgebra.Generic.FreeModule{R}: A OSCAR free module over $\mathbb{Z}$ or $\mathbb{Q}$, it denotes the character group. R is either ZZRingElem or QQFieldElem.

w::Dict{Edge, AbstractAlgebra.Generic.FreeModuleElem{R}}: The axial function

check::Bool=true: Check if the data insered are consistent with those of a GKM graph.

checkLabels::Bool=true: Check if labels are consistent.

Example

Let us construct the GKM graph of the projective line. First of all, we create a graph with two vertices, and one edge.

julia> g = Graph{Undirected}(2)
Undirected graph with 2 nodes and no edges

julia> add_edge!(g, 1, 2);

Let us define our array of labels.

julia> labels = ["a", "b"];

Now, we create the character group. We take a free module of rank 2 over the integers.

julia> M = free_module(ZZ, 2)
Free module of rank 2 over ZZ

We create the axial function. It is a dictionary from the set of edges to the character group. This time we have only one edge.

julia> e = first(edges(g));

julia> w = Dict(e => gens(M)[1] - gens(M)[2])
Dict{Edge, AbstractAlgebra.Generic.FreeModuleElem{ZZRingElem}} with 1 entry:
  Edge(2, 1) => (1, -1)

Finally, we create the GKM graph.

julia> gkm_graph(g, labels, M, w)
GKM graph with 2 nodes, valency 1 and axial function:
b -> a => (1, -1)

Warning

Do not change the number of verices after this.

After you have added all edges using GKMadd_edge!, use GKM_initialize! to calculate the GKM connection (if it is unique) and the curve classes.

GKMtest.empty_gkm_graph — Function

empty_gkm_graph(n::Int64, val::Int64, labels::Vector{String}) -> AbstractGKM_graph

Useful for constructing GKM graph. It return the GKM graph with n fixed points, no edges, valency val and vertices labelled by labels.

julia> G = empty_gkm_graph(2, 2, ["a", "b"])
GKM graph with 2 nodes, valency 0 and axial function:

GKMtest.GKMadd_edge! — Function

GKMadd_edge!(G::AbstractGKM_graph, s::String, d::String, weight::AbstractAlgebra.Generic.FreeModuleElem{R}) where R<:GKM_weight_type

Add an edge to G from the vertex labelled s to the vertex labelled d, the axial function takes value weight in that edge.

Let us construct the same example of gkm_graph, that is the GKM graph of the projective space.

julia> G = empty_gkm_graph(2, 2, ["a", "b"])
GKM graph with 2 nodes, valency 0 and axial function:

julia> wei = gens(G.M)[1] - gens(G.M)[2]
(1, -1)

julia> GKMadd_edge!(G, "b", "a", wei);

julia> G
GKM graph with 2 nodes, valency 1 and axial function:
b -> a => (1, -1)

Warning

Add all edges immediately after creation. Any curve class or cohomology functionality should only be used after all edges have been added. The same holds for GKM_initialize.

GKMadd_edge!(G::AbstractGKM_graph, s::String, d::String, weight::AbstractAlgebra.Generic.FreeModuleElem{R}) where R<:GKM_weight_type

Same as before, but using the number of the vertex instead of the label.

GKMtest.convert_weights — Function

convert_weights(G::AbstractGKM_graph) -> AbstractGKM_graph{QQFieldElem}

It returns $G$, but the character group will be embedded into a free $\mathbb{Q}$-module.

Examples

julia> G_over_Z = generalized_gkm_flag(root_system(:A, 1))
GKM graph with 2 nodes, valency 1 and axial function:
s1 -> id => (-1, 1)

julia> typeof(G_over_Z)
GKMtest.AbstractGKM_graph{ZZRingElem}

julia> G_over_Q = convert_weights(G_over_Z)
GKM graph with 2 nodes, valency 1 and axial function:
s1 -> id => (-1, 1)

julia> typeof(G_over_Q)
GKMtest.AbstractGKM_graph{QQFieldElem}

GKMtest.GKM_isValid — Function

GKM_isValid(gkm::AbstractGKM_graph; printDiagnostics::Bool=true) -> Bool

Return true if the GKM graph is valid. This means:

Every vertex has the same degree

The weights are defined for every edge and every reverse of every edge

The weights belong to the weight lattice

The weights of an edge and its reverse sum to zero

There are the right number of vertex labels

If the valency is at least two, the weights of the graph are 2-independent.

Vertex labels must be unique

The equivariant cohomology ring has rank = number of vertices of graph

The coefficient ring of the equivariant cohomology ring has number of generators = torus rank.